requirements.txt
django==1.5.1
south
django-piston
flup
mysqlclient